+++ CHANGELOG.md
@@ -8,19 +8,27 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Changed
+
+- Made gFTL-shared, yaFyaml, and pFlogger REQUIRED
+
 ### Fixed
 ### Removed
 ### Added
+
+- Added ability for OpenMP and Double Precision to be used with f2py
+  Used by the MAM Optics code
+- Add ability to allow @-symbol to be at beginning or end of sub-repo (still in progress)
+- Emit BASEDIR location during CMake
 
 ## [2.2.2]
 
 ### Added
 
 - Added macro to verify availability of Python modules
-  Use: esma_find_python_module(<module> [REQUIRED])
+  Use: `esma_find_python_module(<module> [REQUIRED])`
 - Added macro to add a post-build check availability of Python modules
-  Use: esma_check_python_module(<module>)
-- Added option is esma_add_library() to use SHARED
+  Use: `esma_check_python_module(<module>)`
+- Added option is `esma_add_library() to use SHARED
 	
 ## [2.2.1]